Initial script for enriching my film scans
This provides the initial rough implementation of a script that adds
EXIF information to my film scans based on an export of entries from
Day One.

I have an iOS Siri Shortcut that I use to record info about a shot I
just took on my cameras. The shortcut exports that info as an entry in
my Day One journal, and previously I manually updated EXIF info in
Lightroom after import -- a particularly tedious task.

This script automates most of that.