Allow option to overwrite files in place
Normally, exiftool(1) copies the original to a file with the suffix
`_original` so that writes are not destructive. However, I probably
want to suppress this, so provide a flag.

Additionally, make the CLI interface a tad bit easier to use by
providing short options, and make the required prefix arg a positional
argument instead.