Translate timestamp to entry's timezone for munged date
Technically, the EXIF timestamp tags such as DateTimeOriginal don't
have a timezone associated with them, so when we populate
DateTimeOriginal and we import the resulting TIFF scan into Lightroom
we end up with incorrect dates and times. Since the timezone
information isn't propagated in the EXIF timestamp, it's best to just
translate them to the timezone local to the entry.