Fall back to local timezone if there's no location metadata
If the entry doesn't have location information, fall back to using the
current local timezone, when calculating the timestamp. A better
option is to propagate the location from the previous entry if it's
fairly close in time (e.g. same day, within an hour maybe) -- but this
is probably good enough for now.