Skip file if frame count if there's no metadata entry
Skip any matching file if there isn't a metadata entry for it. This
would happen when I have a blank scan because either a) I have a
misfire in the middle of the roll and didn't bother to put in an
entry, or b) I accidentally scanned in an empty frame at the end and
didn't delete it.

Either way, just ignore if the frame isn't found in the metadata.