Switch to passing scan files on command file
Instead of using a prefix and a directory then deriving the files to
update from there, accept a list of files to update instead. This
makes it easier to plumb this script elsewhere, plus it makes it
easier to update lab scans where the filenames are likely to be
something more random -- the prefix and directory setup I had here
originally was purely conventional, based on how I name my own scans.