Convert implementation to Rust
Convert my initial Python implementation to Rust to make it possible
to build a statically linked executable that I can move to another
machine quite easily -- for instance so I can install it on my NAS or
on a Raspberry PI. Plus, it's an opportunity for me to learn Rust, so
why not.

This rewrite is a mostly straight port of the existing Python
implementation. I expect it to not be 100% idiomatic Rust.