Replace existing film type entry tags with specific keywords for them
Use 'film:' EXIF keywords to make it obvious that these specify the
type of film used; the bare '120' keyword can be confusing, and using
this prefixed keyword makes it easier to search.

While we're here, refactor when we push the EXIF Keywords tag into the
method where we process all the entry tags -- it's more obvious that's
where it should be.