Overwrite keywords.
exiftool apparently treats `+=` as append-only for array tags and it
applies it for all instances of the tag argument, even if you have any
of that specific tag argument as = -- i.e, if you pass:

  -Foo=Bar -Foo+=Baz

then exiftool will treat all Foo values as being appended to the
existing Keywords tag of the file. Additionally, multiple tag
arguments act as passing an array, so in reality `+=` is unnecessary
strictly speaking, we just have to do:

  -Foo=Bar -Foo=Baz

to set the Foo keyword to the array [Bar,Baz].

For Keywords I want to overwrite, so rerunning scan-metadata on the
same file will have the net effect of applying the exact same tags and
be roughly idempotent. Previously, rerunning on the same file will
re-add the keywords, which end up causing duplicates to show up.