Make output more readable
Make the dryrun output a bit more readable by indenting each argument
and separating each file processed with a newline. Similarly, silence
exiftool since its output of having processed 1 file is pretty noisy
considering we pass only one file at a time to it -- but on the other
hand, we want to see how many of the files we've updated, so count
those and output a final total of files processed.