Improve filename regex for matching frame counts
Improve the regex for deriving the frame count from the filename,
specifically by reducing the number of capture groups and making it
more explicit that we want a numeric suffix. Technically, the previous
implementation could match against a numeric string in the middle of
the stem -- e.g. `foo_0012_bar04` would match but potentially capture on
0012 and not 04.

So, make it a little more strict and only match numeric suffixes.

In the future, this particular regex should be something that I can
pass in -- but that might be too complicated to implement though.