Make EXIF tags be compatible with experimental processor implementation
Refactor the EXIF tag names we use and the values we emit to minimize
the differences between the way tag names and values are handled
between the two EXIF processor implementations.

Additionally, extend the types of values we support even though the
exiftool implementation boils down to treating these values as
strings, as the experimental internal implementation will likely
require calling specific methods depending on the type of the tag
value.