Add more tests and fix timezone lookup bug
My original code for munging the timestamp into the correct timezone
based on the metadata on the entry was actually broken and only
coincidentally worked due to the fact that I've been working on images
in the same local timezone. Specifically, the
parse::<chrono::FixedOffset>() didn't work, as that expects the
timezone to be expressed as an offset from UTC _not_ as a timezone
identifier (e.g. "Americas/Seattle" or "Australia/Sydney"); however,
Day One stores timezone info as a timezone identifier.

Fix the lookup then by adding the chrono-tz crate for handling the
timezone identifier lookup and only fall back to local IFF the TZ
identifier is invalid.

While we're here, do a bit of a cleanup so that ::new() does idiomatic
initialization and calls self.populate_tags(), instead of delegating
that to the caller. This means that callers can treat instances of
MetadataEntry as functionally immutable.