Make filename matching regex even more restrictive
If we have files with filenames like 000021960012 our matcher gets
confused and counts the whole filename as the frame count. So, handle
it by making the regex more restrictive, since in reality we only can
get frame count values up to maybe 42 (for 135 film), and simply
consider the trailing 2 digits of the file.