Use custom serializer for dates
Specifically, explicitly mark creationDate as a DateTime, and handle
parse errors at JSON deserialization time instead of metadata entry
creation time. This makes the error message more explicit -- we'd
complain about the JSON being malformed vs erroring out later on where
it's not obvious that it's the timestamp that's malformed.

This in turn allows me to reuse this bit of serde for handling sunset
and sunrise times later on.