Fix filename matching
Add tests around the filename to metadata matcher, specifically since
I broke it in 30c0cc9d -- the regex matches correctly, but includes
the leading zero; however, we _don't_ include the leading zero for the
first 9 frame counts, and we use a string as the frame count in
MetadataEntry, which means we don't match.

For now, use regexes, but there's probably a refactor hint here to
simply get the trailing digits directly, instead of relying on a
regex.

As an aside, this commit also configures the logger for tests so that
we have logs to make it easier to debug any failed tests.