Swap EXIF implementations
Use the internal Rexiv2-based processor instead of forking
exiftool(1). I've been using the experimental processor as my
"default" for a while now, and I'm confident it works as I expect it
to, so swapping to using it as the actual default; however, leave
exiftool as an option for when I find out specific metadata doesn't
get written to the scans correctly, or as a fallback when I add new
metadata processing.