Handle leader entries when parsing frame counts
I currently don't export the leader entries. However, I'd like to
reuse the MetadataEntry for splitting a larger export into separate
files, one per roll of film, that can be fed into update_scan_exif. To
do this, I'd need to be able to handle the contents of the leader
entry which I create to represent the start of the roll, and which
includes the type of film I've loaded for that camera.

For now, this means refactoring parse_frame_count to use
String::default when we find a non-numeric frame count in the entry
headline, which is what we'd find in the case of the leader entry.

Later on, I'd need to refactor MetadataEntry to be an enumerated type
representing either a leader or a frame; that comes next.