Make split filenames be more descriptive
Add the entry date of the leader entry to the filename of each split
metadata file, to make it easier to see which ZIP file corresponds to
which entries. Keep the index for disambiguation.

NB: It might be better to start the filename with the entry date so
that it's easier to sort; I'll test this out first.