Handle frame counts on half-frame cameras
When recording frame counts for half-frame cameras such as the Olympus
Pen EES-2, the frame count can actually reach 72 or more (36 normal
frames in a standard roll * 2). Since we munge the entry timestamp
with the frame count, and the frame count can be >= 60 in this case,
the resulting munged timestamp is invalid.

To make validly munged timestamps, just use a seconds value that is
the frame count modulo 60.